requirements.txt
fastapi
uvicorn
pillow
python-multipart
loguru
tqdm